+++ tests/integration/distribution/keeper/msg_server_test.go
@@ -108,7 +108,7 @@ func initFixture(t *testing.T) *fixture {
 	require.NoError(t, stakingKeeper.Params.Set(newCtx, stakingtypes.DefaultParams()))
 
 	poolKeeper := poolkeeper.NewKeeper(
-		cdc, runtime.NewKVStoreService(keys[pooltypes.StoreKey]), accountKeeper, bankKeeper, authority.String(),
+		cdc, runtime.NewKVStoreService(keys[pooltypes.StoreKey]), accountKeeper, bankKeeper, stakingKeeper, authority.String(),
 	)
 
 	distrKeeper := distrkeeper.NewKeeper(